+++ Testing/src/main/java/com/mopub/mobileads/testing/NativeListViewFragment.java
@@ -14,10 +14,10 @@ import android.widget.ListView;
 import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 
-import com.mopub.mobileads.mbridge.MBridgeAdRenderer;
 import com.mopub.nativeads.FacebookAdRenderer;
 import com.mopub.nativeads.GooglePlayServicesAdRenderer;
 import com.mopub.nativeads.GooglePlayServicesViewBinder;
+import com.mopub.nativeads.MintegralAdRenderer;
 import com.mopub.nativeads.MoPubAdAdapter;
 import com.mopub.nativeads.MoPubStaticNativeAdRenderer;
 import com.mopub.nativeads.PangleAdRenderer;
@@ -136,9 +136,10 @@ public class NativeListViewFragment extends Fragment {
                         .callToActionId(R.id.native_cta)
                         .adChoicesRelativeLayoutId(R.id.native_privacy_information_icon_layout)
                         .build());
-        // Set up a renderer for MBridge ads.
-        final MBridgeAdRenderer mBridgeAdRenderer = new MBridgeAdRenderer(
-                new MBridgeAdRenderer.ViewBinder.Builder(R.layout.native_ad_mbridge_list_item)
+
+        // Set up a renderer for Mintegral ads
+        final MintegralAdRenderer mintegralAdRenderer = new MintegralAdRenderer(
+                new MintegralAdRenderer.ViewBinder.Builder(R.layout.native_ad_list_item)
                         .titleId(R.id.native_title)
                         .textId(R.id.native_text)
                         .mediaViewId(R.id.native_main_image)
@@ -146,6 +147,7 @@ public class NativeListViewFragment extends Fragment {
                         .callToActionId(R.id.native_cta)
                         .adChoicesId(R.id.native_privacy_information_icon_image)
                         .build());
+
         // Register the renderers with the MoPubAdAdapter and then set the adapter on the ListView.
         // The first renderer that can handle a particular native ad gets used.
         // We are prioritizing network renderers.
@@ -155,7 +157,7 @@ public class NativeListViewFragment extends Fragment {
         mAdAdapter.registerAdRenderer(facebookAdRenderer);
         mAdAdapter.registerAdRenderer(staticAdRender);
         mAdAdapter.registerAdRenderer(pangleAdRenderer);
-        mAdAdapter.registerAdRenderer(mBridgeAdRenderer);
+        mAdAdapter.registerAdRenderer(mintegralAdRenderer);
         listView.setAdapter(mAdAdapter);
 
         updateRequestParameters(views);