+++ Testing/src/main/java/com/mopub/mobileads/testing/NativeManualFragment.java
@@ -15,11 +15,11 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 
-import com.mopub.mobileads.mbridge.MBridgeAdRenderer;
 import com.mopub.nativeads.AdapterHelper;
 import com.mopub.nativeads.FacebookAdRenderer;
 import com.mopub.nativeads.GooglePlayServicesAdRenderer;
 import com.mopub.nativeads.GooglePlayServicesViewBinder;
+import com.mopub.nativeads.MintegralAdRenderer;
 import com.mopub.nativeads.MoPubNative;
 import com.mopub.nativeads.MoPubStaticNativeAdRenderer;
 import com.mopub.nativeads.NativeAd;
@@ -185,9 +185,10 @@ public class NativeManualFragment extends Fragment {
                         .callToActionId(R.id.native_cta)
                         .adChoicesRelativeLayoutId(R.id.native_privacy_information_icon_layout)
                         .build());
-        // Set up a renderer for MBridge ads.
-        final MBridgeAdRenderer mBridgeAdRenderer = new MBridgeAdRenderer(
-                new MBridgeAdRenderer.ViewBinder.Builder(R.layout.native_ad_mbridge_list_item)
+
+        // Set up a renderer for Mintegral ads
+        final MintegralAdRenderer mintegralAdRenderer = new MintegralAdRenderer(
+                new MintegralAdRenderer.ViewBinder.Builder(R.layout.native_ad_list_item)
                         .titleId(R.id.native_title)
                         .textId(R.id.native_text)
                         .mediaViewId(R.id.native_main_image)
@@ -195,16 +196,16 @@ public class NativeManualFragment extends Fragment {
                         .callToActionId(R.id.native_cta)
                         .adChoicesId(R.id.native_privacy_information_icon_image)
                         .build());
+
         // The first renderer that can handle a particular native ad gets used.
         // We are prioritizing network renderers.
-
         mMoPubNative.registerAdRenderer(facebookAdRenderer);
         mMoPubNative.registerAdRenderer(googlePlayServicesAdRenderer);
         mMoPubNative.registerAdRenderer(verizonNativeAdRenderer);
         mMoPubNative.registerAdRenderer(pangleAdRenderer);
         mMoPubNative.registerAdRenderer(referenceNativeAdRenderer);
         mMoPubNative.registerAdRenderer(moPubStaticNativeAdRenderer);
-        mMoPubNative.registerAdRenderer(mBridgeAdRenderer);
+        mMoPubNative.registerAdRenderer(mintegralAdRenderer);
 
         mMoPubNative.makeRequest(mRequestParameters);