+++ Mintegral/CHANGELOG.md
@@ -1,5 +1,3 @@
 ## Changelog
-
 * 15.5.11.0
-    * release time: 2021-05-06
-    * This version of the adapters has been certified with mBridge v_15.5.11 and MoPub SDK 5.16.4.
+    * This version of the adapters has been certified with Mintegral 15.5.11 and MoPub SDK 5.17.0.